Lagrer nyXML og linker til filen i footer

+++ hentdata.php
@@ -17,6 +17,11 @@ if (!$data) {
 }
 
 $arr = json_decode($data,true); //Gjør dataen om til array
+
+if(array_key_exists('-1', $arr['query']['pages'])){
+	exit('Dette stedet har ikke en Wikipedia-artikkel. <a href="https://no.wikipedia.org/wiki/'.$stedsnavn.'"">Kanskje du kan lage en?</a>');
+}
+
 $output_details = reset($arr['query']['pages']); //Velger første element da id er dynamisk. http://php.net/manual/en/function.reset.php
 echo $output_details['extract']; // Henter ut teksten
 echo "<br><a href='https://no.wikipedia.org/wiki/".$stedsnavn."'>Les mer på Wikipedia</a>";
